requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
pydantic
numpy
torch
pandas
python-docx
openpyxl
pypdf
requests
sqlalchemy
pytest
httpx

# OCR dependencies
pytesseract>=0.3.10
pdf2image>=1.17.0
easyocr>=1.7.0  # GPU-accelerated OCR (GPU 2)
pillow>=10.0.0
opencv-python-headless>=4.8.0
rank-bm25==0.2.2  # For hybrid search (Dense + BM25)
